rename jspreadsheet into spreadsheet-editor
