+++ builder.js
@@ -29,8 +29,8 @@ module.exports = (env) => {
         './src/ts/ketcher-editor.jsx',
         './src/ts/compounds-table.jsx',
         './src/ts/users-table.jsx',
-        './src/ts/jspreadsheet.jsx',
-        './src/ts/jspreadsheet.utils.ts',
+        './src/ts/spreadsheet-editor.jsx',
+        './src/ts/spreadsheet.utils.ts',
         './src/ts/tags.ts',
         './src/ts/admin.ts',
         './src/ts/profile.ts',
