+++ src/templates/edit.html
@@ -116,7 +116,7 @@
 
 {% include 'steps-edit.html' %}
 
-{% include 'jspreadsheet.html' %}
+{% include 'spreadsheet-editor.html' %}
 
 {% include 'links.html' %}
