+++ src/templates/spreadsheet-editor.html
@@ -1,0 +1,9 @@
+<section id='spreadsheetEditor' aria-label='{{ 'Spreadsheet Editor'|trans }}'>
+  <h3 data-action='toggle-next' data-toggle-target='spreadsheetEditorDiv' data-opened-icon='fa-caret-down' data-closed-icon='fa-caret-right' class='d-inline togglable-section-title mt-2' tabindex='0' role='button' aria-expanded='false' aria-controls='spreadsheetEditorDiv'>
+    <i id='sheetEditorIcon' class='fas fa-caret-right fa-fw mr-2'></i>{{ 'Spreadsheet Editor'|trans }}
+  </h3>
+  <div id='spreadsheetEditorDiv' class='m-2' hidden data-save-hidden='spreadsheetEditorDiv'>
+    <div id='spreadsheet-importer-root'></div>
+  </div>
+</section>
+<hr>
