+++ src/templates/view.html
@@ -144,7 +144,7 @@
 
 {% include 'steps-view.html' %}
 
-{% include 'jspreadsheet.html' %}
+{% include 'spreadsheet-editor.html' %}
 
 {% include 'links.html' %}
