+++ src/ts/spreadsheet.utils.ts
